+++ src/InteractsWithForgeAll.php
@@ -1,0 +1,7 @@
+<?php
+namespace LKDevelopment\ForgeConnect;
+
+trait InteractsWithForgeAll
+{
+    use InteractsWithForgeConnectDir, InteractsWithForgeConfiguration, InteractsWithForgeCache;
+}