+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
-
+## 1.1.0 - 2016-12-03
+- Your Forge Password will be encrypted with AES
+- Added Caching to servers:list and server:connect to save your servers on your local disk (Cachelifetime 3600 sec (1h))
+- Code Refactoring
 ## 1.0.0 - 2016-12-01
 - Initial Release